Added Section Components & Automatic Labeling of Sections
This commit adds the section component which is used to structure the page and display content as well as the section-group component (you can find more details on it in the corresponding *.twig files). I might abstractify the section component in the future (i.e. remove the necessity for a title).