Dropdowns Now Optional
When you don't define the "dropdown" attribute for a nav item, no dropdown will appear. You can still give it a link by providing a "link" attribute. If both a dropdown and a link attribute are defined, the link attribute will override the dropdown attribute but the dropdown will still appear with all the links in place.