Improved News Explorer & Navigation Bar
Several changes were made in this commit:

An alternative text is now shown in the news explorer when there are no search results.

In the database, the "products" table was renamed to "news_categories".

The jobs category is not included in the news dropdown of the navbar. Instead it has its own nav bar menu item.