Finished Implementation of the Landing Page
The landing page now shows up when a file called "index.html" is placed inside the "landingpage" folder.

The landing page is being displayed in an iFrame, that's why relative paths work inside the landing page file.

The dimensions of the landing page window are controlled by the x-parent-width and x-parent-height meta tags.